feat: put handle navigate to question input into a separate hook
