+++ src/Hooks/index.ts
@@ -8,9 +8,9 @@ export { default as useShareBotUrl } from "./useShareBotUrl/useShareBotUrl";
 export { default as useMainButtonHandler } from "./useMainButtonHandler/useMainButtonHandler";
 export { default as useSaveSpreadState } from "./useSaveSpreadState/useSaveSpreadState";
 export { default as useNavigation } from "./useNavigation/useNavigation";
-export { default as useRewardPopup } from "./useRewardPopup/useRewardPopup";
 export { default as useInfoPopup } from "./useInfoPopup/useInfoPopup";
 export { default as useErrorPopup } from "./useErrorPopup/useErrorPopup";
 export { default as useSettingsButton } from "./useSettingsButton/useSettingsButton";
 export { default as useReadings } from "./useReadings/useReadings";
 export { default as useMainButtonState } from "./useMainButtonState/useMainButtonState";
+export { default as useNavigateToQuestion } from "./useNavigateToQuestion/useNavigateToQuestion";
